// packages/discord/src/operations/getMessage.ts
export const GetMessageInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
  channel_id: Schema.String.pipe(T.PathParam()),
  message_id: Schema.String.pipe(T.PathParam()),
}).pipe(
  T.Http({
    method: "GET",
    path: "/channels/{channel_id}/messages/{message_id}",
  }),
);
export type GetMessageInput = typeof GetMessageInput.Type;
// ...
export type GetMessageOutput = typeof GetMessageOutput.Type;

// The operation
export const getMessage = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: GetMessageInput,
  outputSchema: GetMessageOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));